+++ package.json
@@ -10,6 +10,7 @@
     "@okta/okta-signin-widget": "^3.7.3",
     "@reduxjs/toolkit": "^1.2.5",
     "@sentry/browser": "^5.12.4",
+    "axios": "^0.19.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-redux": "^7.2.0",
