+++ src/App.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
-import Login from './components/login/LoginForm'
+import LoginForm from './components/login/LoginForm'
+
 import { Security, SecureRoute, ImplicitCallback } from '@okta/okta-react'
-import Home from './components/dashboard/Home'
 import SignUp from './components/signUpForm/SignUpOverlay'
 import { CssBaseline } from '@material-ui/core'
+import AdminDash from './components/AdminDash/AdminDashboard'
 
 function onAuthRequired({ history }) {
   history.push('/login')
@@ -20,12 +21,12 @@ function App() {
         onAuthRequired={onAuthRequired}
         pkce={true}
       >
-        <Route path='/' exact={true} component={Home} />
         <Route
           path='/login'
-          render={() => <Login baseUrl='https://dev-529730.okta.com' />}
+          render={() => <LoginForm baseUrl='https://dev-529730.okta.com' />}
         />
         <Route exact path='/signup' component={SignUp} />
+        <SecureRoute exact path='/' component={AdminDash} />
         <Route path='/implicit/callback' component={ImplicitCallback} />
       </Security>
     </>
