--- src/components/dashboard/Home.js
@@ -1,55 +1,0 @@
-import React, { Component } from 'react'
-import { Link } from 'react-router-dom'
-import { withAuth } from '@okta/okta-react'
-
-export default withAuth(
-  class Home extends Component {
-    constructor(props) {
-      super(props)
-      this.state = { authenticated: null }
-      this.checkAuthentication = this.checkAuthentication.bind(this)
-      this.checkAuthentication()
-    }
-
-    async checkAuthentication() {
-      const authenticated = await this.props.auth.isAuthenticated()
-      if (authenticated !== this.state.authenticated) {
-        this.setState({ authenticated })
-      }
-    }
-
-    componentDidUpdate() {
-      this.checkAuthentication()
-    }
-
-    render() {
-      if (this.state.authenticated === null) return null
-
-      const button = this.state.authenticated ? (
-        <button
-          onClick={() => {
-            this.props.auth.logout()
-          }}
-        >
-          Logout
-        </button>
-      ) : (
-        <button
-          onClick={() => {
-            this.props.auth.login()
-          }}
-        >
-          Sign In
-        </button>
-      )
-
-      return (
-        <div>
-          <Link to='/'>Home</Link>
-          <br />
-          {button}
-        </div>
-      )
-    }
-  }
-)
